requirements.txt
peewee~=3.17.5
python-dotenv~=1.0.1
python-telegram-bot==21.4
requests~=2.32.3